renamed files, removed repeating tests
