+++ code/TS1_DivisionByZero/StaticField/c++/StaticField.cpp
