+++ code/TS3_NullPointerException/Native/java/InferNativeNPETests.java
