+++ code/TS3_NullPointerException/Trivial/java/InferTrivialNPETests.java
