+++ code/TS3_NullPointerException/WithStandardLib/java/InferSTDLibNPETests.java
